Mastering

+++ app/Http/Controllers/UserController.php
@@ -49,6 +49,12 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate([
+            'name' => 'required',
+            'email' => 'email|required',
+            'password' => 'required',
+        ]);
+
         if (auth()->user()->isAdmin()){
             $user = new User;
             $user->name = $request->name;
@@ -137,9 +143,13 @@ class UserController extends Controller
     public function destroy($id)
     {
         if (auth()->user()->isAdmin()){
-            $user = User::findOrFail($id);
-            $user->delete();
-            return redirect('/user/admin');
+            if (auth()->user()->id != $id){
+                $user = User::findOrFail($id);
+                $user->delete();
+                return redirect('/user/admin');
+            } else {
+                return redirect('/user/admin');
+            }
         } else {
             abort(404);
         }
